Name organizing and readme update
