+++ Plugins/Controls/ChannelCreator.csproj
@@ -11,7 +11,7 @@
     <ItemGroup>
       <PluginFiles Include="$(TargetDir)$(TargetName).dll" />
     </ItemGroup>
-    <Message Text="Copying @(PluginFiles) to $(SolutionDir)TS3AudioBot\bin\Debug\netcoreapp3.1\Plugins\" Importance="high" />
-    <Copy SourceFiles="@(PluginFiles)" DestinationFolder="$(SolutionDir)TS3AudioBot\bin\Debug\netcoreapp3.1\Plugins\" />
+    <Message Text="Copying @(PluginFiles) to $(SolutionDir)TS3AudioBot\bin\Debug\net6.0\Plugins\" Importance="high" />
+    <Copy SourceFiles="@(PluginFiles)" DestinationFolder="$(SolutionDir)TS3AudioBot\bin\Debug\net6.0\Plugins\" />
   </Target>
 </Project>
